Test runs without bulletproof

+++ athena/src/main/java/project/athena/Athena.java
@@ -10,5 +10,5 @@ public interface Athena {
     RegisterStruct Register(PK_Vector pkv);
     Ballot Vote(CredentialTuple dv, PK_Vector pkv, int vote, int cnt, int nc);
     TallyStruct Tally(SK_Vector skv,  int nc);
-    boolean Verify(PK_Vector pkv, int nc,  Map<BigInteger, Integer> b, PFStruct pf);
+    boolean Verify(PK_Vector pkv, int nc,  Map<Integer, Integer> b, PFStruct pf);
 }
